提交

+++ aab/apk2aab.py
@@ -2,10 +2,9 @@ import os
 import shutil
 import zipfile
 
-import apkCompile as apk
-
-apkPath = './resource/app.apk'
 outPath = './resource/app'
+
+apktoolPath = '../tools/apktool_2.6.1.jar'
 
 resource = './resource'
 aapt2 = '../tools/aapt2'
@@ -15,10 +14,6 @@ compiled_resources = './resource/compiled_resources.zip'
 android_jar = '../tools/android_31.jar'
 baseApk = f'{resource}/base.apk'
 baseApkDir = f'{resource}/base'
-
-jks = f'{resource}/ping.jks'
-password = "oryuauzfp"
-alias = "pahyctxfp"
 
 
 def compileRes():
@@ -31,6 +26,11 @@ def linkRes():
           '--min-sdk-version 19 --target-sdk-version 31 --version-code 1 --version-name 1.0 ' \
           f'--manifest {outPath}/AndroidManifest.xml ' \
           f'-R {compiled_resources} --auto-add-overlay'
+    os.system(cmd)
+
+
+def decodeDex(apkPath, outputPath):
+    cmd = f"java -jar {apktoolPath} d {apkPath} -s -f -o {outputPath}"
     os.system(cmd)
 
 
@@ -55,8 +55,9 @@ def moveResuoce():
     shutil.copytree(f'{outPath}/lib', f"{baseApkDir}/lib/", dirs_exist_ok=True)
 
     # copy unkwon
-    shutil.copytree(f'{outPath}/unknown', f'{baseApkDir}/unknown/', dirs_exist_ok=True)
-    shutil.move(f'{baseApkDir}/unknown', f'{baseApkDir}/root')
+    if os.path.exists(f'{outPath}/unknown'):
+        shutil.copytree(f'{outPath}/unknown', f'{baseApkDir}/unknown/', dirs_exist_ok=True)
+        shutil.move(f'{baseApkDir}/unknown', f'{baseApkDir}/root')
 
     # copy kotlin
     shutil.copytree(f'{outPath}/kotlin', f'{baseApkDir}/root/kotlin/', dirs_exist_ok=True)
@@ -99,15 +100,15 @@ def compileAAB():
     os.system(cmd)
 
 
-def signAAB():
+def signAAB(jks, password, alias, outputaab):
     # cmd = f'{apksigner} sign --ks {jks} --ks-key-alias {alias} --ks-pass pass:{password} --in {resource}/base.aab --out {resource}/base.apks'
-
-    cmd = f'jarsigner -digestalg SHA1 -sigalg SHA1withRSA -keystore {jks} -storepass {password} -keypass {password} {resource}/base.aab {alias}'
+    cmd = f'jarsigner -verbose -sigalg SHA256withRSA -digestalg SHA-256 -keystore {jks} -storepass {password} -keypass {password} -signedjar {outputaab} {resource}/base.aab {alias}'
+    # cmd = f'jarsigner -digestalg SHA1 -sigalg SHA1withRSA -keystore {jks} -storepass {password} -keypass {password} {resource}/base.aab {alias}'
     os.system(cmd)
 
 
-def installAAB():
-    cmd = f'java -jar {bundletool} build-apks --bundle={resource}/base.aab --output={resource}/base.apks --ks={jks} --ks-pass=pass:{password} --ks-key-alias={alias} --key-pass=pass:{password}'
+def installAAB(aab, jks, password, alias):
+    cmd = f'java -jar {bundletool} build-apks --bundle={aab} --output={resource}/base.apks --ks={jks} --ks-pass=pass:{password} --ks-key-alias={alias} --key-pass=pass:{password}'
     os.system(cmd)
 
     # install apks
@@ -115,7 +116,7 @@ def installAAB():
     os.system(cmd)
 
 
-def aab2Apks(aabPath):
+def aab2Apks(aabPath, jks, password, alias):
     apks = f'{resource}/output/app.apks'
     unzipDir = f'{resource}/output/app/'
     cmd = f'java -jar {bundletool} build-apks --mode=universal --bundle={aabPath} --output={apks} --ks={jks} --ks-pass=pass:{password} --ks-key-alias={alias} --key-pass=pass:{password}'
@@ -125,17 +126,26 @@ def aab2Apks(aabPath):
     os.system(cmd)
 
 
-def start():
-    apk.decodeDex(apkPath, outPath)
+def start(apk, jks, password, alias, outputaab):
+    if os.path.exists(resource):
+        shutil.rmtree(resource)
+
+    decodeDex(apk, outPath)
     compileRes()
     linkRes()
     moveResuoce()
     zipBaseApk()
     compileAAB()
-    signAAB()
-    installAAB()
+    signAAB(jks, password, alias, outputaab)
+    installAAB(outputaab, jks, password, alias)
+
 
 if __name__ == '__main__':
-    start()
-    # aab2Apks(f'{resource}/base.aab')
+    apkPath = './com.fruit.ezdi.slot.apk'
+    jks = f'./com.unnu.happy.slots.jks'
+    password = "flowerwild"
+    alias = "wildflower"
+    outputaab = 'com.fruit.ezdi.slot.aab'
+    start(apkPath, jks, password, alias, outputaab)
 
+    # aab2Apks(f'{resource}/signed.aab')
